requirements.txt
stable-baselines3[extra]    
gymnasium                   
numpy 
pandas                  
pyzmq                          
fastapi 
uvicorn 
websockets    
